fix: build internal/config on Windows
internal/config/config.go used syscall.Flock and syscall.LOCK_EX/LOCK_UN
directly, so `go install github.com/marcus/td@latest` failed on Windows
with `undefined: syscall.Flock`.

Split withConfigLock into platform-specific files, mirroring the existing
pattern in internal/db/lock_{unix,windows}.go and
internal/serve/portfile_{unix,windows}.go:

  - lock_unix.go    -> flock-based (unchanged behavior)
  - lock_windows.go -> LockFileEx with LOCKFILE_EXCLUSIVE_LOCK (blocking,
                       matching the Unix LOCK_EX semantics)

All locking-relevant config tests pass on Windows, including the
TestEdgeCases/concurrent_operations stress test that drives ten goroutines
through SetFocus/SetActiveWorkSession concurrently. The two
TestPermissionErrors failures are pre-existing (the test uses POSIX
chmod 0000/0555, which don't map to Windows ACLs) and unrelated to this
change.
